+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 ## Unreleased
 
+## 13.0.0
+
+- `input.validate` is of type ValidationState
+- `<input name="foo" required oninput="setCustomValidity('')" oninvalid="setCustomValidity('Required')">` now executes attribute event handlers
+
 ## 12.1.0
 
 - add `required` property to input, select, and textarea